refactor: put program machine in a separate file
