+++ .vscode/settings.json
@@ -4,5 +4,6 @@
   "i18n-ally.sortKeys": true,
   "cSpell.words": [
     "Vitesse"
-  ]
+  ],
+  "editor.tabSize": 2
 }
