+++ src/pages/programs/[id].vue
@@ -126,259 +126,14 @@
 
 <script lang="ts">
 import { defineComponent, computed, watch } from 'vue'
-import { createMachine } from 'xstate'
 import { useMachine } from '@xstate/vue'
 import ChartBarIcon from '/@vite-icons/heroicons-outline/chart-bar.vue'
 import ChipIcon from '/@vite-icons/heroicons-outline/chip.vue'
 import IdentificationIcon from '/@vite-icons/heroicons-outline/identification.vue'
 
-import { ProgramState, useProgram } from '/~/composables/programs'
-
-enum AvailableAction {
-  start = 'start',
-  stop = 'stop',
-  restart = 'restart',
-  modify = 'modify',
-}
-
-enum ProgramMachineActions {
-  START = 'START',
-  STOP = 'STOP',
-  RESTART = 'RESTART',
-  MODIFY = 'MODIFY',
-}
-
-interface ProgramMachineMeta {
-  actions: AvailableAction[]
-}
-
-interface ProgramMachineContext {}
-
-type ProgramMachineState =
-  | {
-    value: 'SELECTING'
-    context: ProgramMachineContext
-  }
-  | {
-    value: ProgramState.STARTING
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.RUNNING
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.STOPPING
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.STOPPED
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.BACKOFF
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.EXITED
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.FATAL
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-  | {
-    value: ProgramState.UNKNOWN
-    context: ProgramMachineContext
-    meta: ProgramMachineMeta
-  }
-
-type ProgramMachineEvent =
-  | {
-    type: AvailableAction.start
-  }
-  | {
-    type: AvailableAction.stop
-  }
-  | {
-    type: AvailableAction.restart
-  }
-  | {
-    type: AvailableAction.modify
-  }
-  | {
-    type: ProgramState
-  }
-
-const programMachine = createMachine<ProgramMachineContext, ProgramMachineEvent, ProgramMachineState>({
-  id: 'program',
-  initial: ProgramState.UNKNOWN,
-  states: {
-    [ProgramState.STARTING]: {
-      meta: {
-        actions: [
-          AvailableAction.stop,
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.stop]: {
-          actions: ProgramMachineActions.STOP,
-        },
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.RUNNING]: {
-      meta: {
-        actions: [
-          AvailableAction.restart,
-          AvailableAction.stop,
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.restart]: {
-          actions: ProgramMachineActions.RESTART,
-        },
-        [AvailableAction.stop]: {
-          actions: ProgramMachineActions.STOP,
-        },
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.STOPPING]: {
-      meta: {
-        actions: [
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.STOPPED]: {
-      meta: {
-        actions: [
-          AvailableAction.start,
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.start]: {
-          actions: ProgramMachineActions.START,
-        },
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.BACKOFF]: {
-      meta: {
-        actions: [
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.EXITED]: {
-      meta: {
-        actions: [
-          AvailableAction.start,
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.start]: {
-          actions: ProgramMachineActions.START,
-        },
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.FATAL]: {
-      meta: {
-        actions: [
-          AvailableAction.start,
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.start]: {
-          actions: ProgramMachineActions.START,
-        },
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-
-    [ProgramState.UNKNOWN]: {
-      meta: {
-        actions: [
-          AvailableAction.modify,
-        ],
-      },
-
-      on: {
-        [AvailableAction.modify]: {
-          actions: ProgramMachineActions.MODIFY,
-        },
-      },
-    },
-  },
-
-  on: {
-    [ProgramState.BACKOFF]: ProgramState.BACKOFF,
-    [ProgramState.RUNNING]: ProgramState.RUNNING,
-    [ProgramState.STOPPING]: ProgramState.STOPPING,
-    [ProgramState.STOPPED]: ProgramState.STOPPED,
-    [ProgramState.EXITED]: ProgramState.EXITED,
-    [ProgramState.FATAL]: ProgramState.FATAL,
-    [ProgramState.UNKNOWN]: ProgramState.UNKNOWN,
-  },
-})
-
-function mergeMeta<T>(meta: Record<string, T>): T {
-  return Object.keys(meta).reduce((acc, key) => {
-    const value = meta[key]
-
-    // Assuming each meta value is an object
-    Object.assign(acc, value)
-
-    return acc
-  }, {} as T)
-}
+import { useProgram } from '/~/composables/programs'
+import { programMachine, ProgramMachineActions, ProgramMachineMeta } from '/~/machines/program'
+import { mergeMeta } from '/~/machines/utils'
 
 export default defineComponent({
   props: {
